+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-05-14 18:30:06
-// Commit: feat(core): add async core stream provider
+// Updated at: 2026-05-14 18:55:13
+// Commit: feat(core): implement responsive core view component
 
 class coreComponentState {
-  final String commitRef = '1156';
-  final String lastMessage = 'feat(core): add async core stream provider';
-  final DateTime timestamp = DateTime.parse('2026-05-14T18:30:06+05:00');
+  final String commitRef = '1157';
+  final String lastMessage = 'feat(core): implement responsive core view component';
+  final DateTime timestamp = DateTime.parse('2026-05-14T18:55:13+05:00');
 }